remove intervention policy group form.
